flask-api/requirements.txt
Flask
pandas
numpy
scikit-learn
joblib
Flask-Cors
selenium
Pillow
easyocr
beautifulsoup4
requests
